Reorganize sources into directories, clean up includes
