+++ include/noarr/structures.hpp
@@ -1,10 +1,14 @@
 /**
  * @file structures.hpp
- * @brief this file aggregates low-level includes for the entire "structures" module
+ * @brief This file aggregates the definitions necessary for all programs using noarr structures. They allow defining data layouts (including custom ones), querying their basic properties, and calculating sizes and offsets.
  */
 
-#include "structures/layouts.hpp"
-#include "structures/views.hpp"
-#include "structures/slice.hpp"
-#include "structures/funcs.hpp"
-#include "structures/wrapper.hpp"
+#include "structures/base/contain.hpp"
+#include "structures/base/signature.hpp"
+#include "structures/base/state.hpp"
+#include "structures/base/structs_common.hpp"
+#include "structures/base/utility.hpp"
+#include "structures/structs/layouts.hpp"
+#include "structures/structs/scalar.hpp"
+#include "structures/extra/funcs.hpp"
+#include "structures/extra/struct_traits.hpp"
