+++ include/noarr/structures/base/signature.hpp
