+++ include/noarr/structures/base/structs_common.hpp
