+++ include/noarr/structures/base/utility.hpp
