+++ include/noarr/structures/extra/struct_traits.hpp
@@ -1,10 +1,9 @@
 #ifndef NOARR_STRUCTURES_STRUCT_TRAITS_HPP
 #define NOARR_STRUCTURES_STRUCT_TRAITS_HPP
 
-#include <type_traits>
-
-#include "signature.hpp"
-#include "state.hpp"
+#include "../base/signature.hpp"
+#include "../base/state.hpp"
+#include "../base/utility.hpp"
 
 namespace noarr {
