+++ include/noarr/structures/interop/bag.hpp
@@ -4,7 +4,7 @@
 #include <memory>
 #include <vector>
 
-#include "wrapper.hpp"
+#include "../extra/wrapper.hpp"
 
 namespace noarr {
