+++ include/noarr/structures/interop/serialize_data.hpp
@@ -4,7 +4,7 @@
 #include <istream>
 #include <ostream>
 
-#include "traverser.hpp"
+#include "../extra/traverser.hpp"
 
 namespace noarr {
