+++ include/noarr/structures/structs/scalar.hpp
@@ -1,9 +1,11 @@
 #ifndef NOARR_STRUCTURES_SCALAR_HPP
 #define NOARR_STRUCTURES_SCALAR_HPP
 
-#include "structs_common.hpp"
-#include "contain.hpp"
-#include "signature.hpp"
+#include "../base/contain.hpp"
+#include "../base/signature.hpp"
+#include "../base/state.hpp"
+#include "../base/structs_common.hpp"
+#include "../base/utility.hpp"
 
 namespace noarr {
