+++ include/noarr/structures/structs/views.hpp
@@ -1,8 +1,11 @@
 #ifndef NOARR_STRUCTURES_VIEWS_HPP
 #define NOARR_STRUCTURES_VIEWS_HPP
 
-#include "structs_common.hpp"
-#include "state.hpp"
+#include "../base/contain.hpp"
+#include "../base/signature.hpp"
+#include "../base/state.hpp"
+#include "../base/structs_common.hpp"
+#include "../base/utility.hpp"
 
 namespace noarr {
