+++ tests/containers_pipes_test.cpp
@@ -4,7 +4,7 @@
 #include <iostream>
 #include <tuple>
 
-#include "noarr/structures_extended.hpp"
+#include <noarr/structures_extended.hpp>
 #include "noarr_test_defs.hpp"
 
 TEST_CASE("Pipes vector", "[resizing]")
