+++ tests/iterator_test.cpp
@@ -9,7 +9,8 @@
 #include <array>
 #include <iostream>
 
-#include "noarr/structures_extended.hpp"
+#include <noarr/structures.hpp>
+#include <noarr/structures/extra/iterator.hpp>
 
 TEST_CASE("Iterator trivial", "[iterator]") {
 	auto array = noarr::array<'x', 20, noarr::scalar<int>>();
