+++ tests/serialize_data_test.cpp
@@ -2,8 +2,8 @@
 
 #include <sstream>
 
-#include "noarr/structures_extended.hpp"
-#include "noarr/structures/serialize_data.hpp"
+#include <noarr/structures_extended.hpp>
+#include <noarr/structures/interop/serialize_data.hpp>
 
 TEST_CASE("Deserialize data", "[serialize_data]") {
 	noarr::array<'x', 3, noarr::array<'y', 3, noarr::scalar<int>>> structure;
