+++ tests/size_dots_test.cpp
@@ -3,7 +3,8 @@
 #include <iostream>
 #include <array>
 
-#include "noarr/structures_extended.hpp"
+#include <noarr/structures.hpp>
+#include <noarr/structures/extra/wrapper.hpp>
 #include "noarr_test_defs.hpp"
 
 using noarr::idx;
